+++ stm32/lib/pcap02/src/pcap02.c
@@ -2,6 +2,8 @@
 
 #include "board.h"
 #include "i2c.h"
+#include "sensor.h"
+#include "sensors.h"
 #include "userConfig.h"
 
 // Private Globals
@@ -802,7 +804,7 @@ size_t pcap02_measure_capacitance(pcap02_result_t *result) {
   return 1;
 }
 
-size_t pcap02_measure(uint8_t *data, SysTime_t ts) {
+size_t pcap02_measure(uint8_t *data, SysTime_t ts, uint32_t idx) {
   pcap02_result_t result;
 
   // read sensor
@@ -812,10 +814,20 @@ size_t pcap02_measure(uint8_t *data, SysTime_t ts) {
 
   const UserConfiguration *cfg = UserConfigGet();
 
-  // encode measurement
-  size_t data_len = EncodePCAP02Measurement(
-      ts.Seconds, cfg->logger_id, cfg->cell_id,
-      PCAP02_REFERENCE_CAPACITOR_PF * fixed_to_double(&result), data);
+  // metadata
+  Metadata meta = Metadata_init_zero;
+  meta.ts = ts.Seconds;
+  meta.logger_id = cfg->logger_id;
+  meta.cell_id = cfg->cell_id;
+
+  // encode measuremnet
+  size_t data_len = 0;
+  double cap = PCAP02_REFERENCE_CAPACITOR_PF * fixed_to_double(&result);
+  SensorStatus status = EncodeDoubleMeasurement(
+      meta, cap, SensorType_PCAP02_CAPACITANCE, data, &data_len);
+  if (status != SENSOR_OK) {
+    return -1;
+  }
 
   return data_len;
 }