+++ stm32/lib/sdi12/src/teros21.c
@@ -1,5 +1,7 @@
 #include "teros21.h"
 
+#include "sensor.h"
+#include "sensors.h"
 #include "userConfig.h"
 
 SDI12Status Teros21ParseMeasurement(const char *buffer, Teros21Data *data) {
@@ -45,7 +47,7 @@ SDI12Status Teros21GetMeasurement(char addr, Teros21Data *data) {
   return status;
 }
 
-size_t Teros21Measure(uint8_t *data, SysTime_t ts) {
+size_t Teros21Measure(uint8_t *data, SysTime_t ts, uint32_t idx) {
   Teros21Data sens_data = {};
   SDI12Status status = Teros21GetMeasurement('0', &sens_data);
   if (status != SDI12_OK) {
@@ -54,9 +56,30 @@ size_t Teros21Measure(uint8_t *data, SysTime_t ts) {
 
   const UserConfiguration *cfg = UserConfigGet();
 
-  size_t data_len =
-      EncodeTeros21Measurement(ts.Seconds, cfg->logger_id, cfg->cell_id,
-                               sens_data.matric_pot, sens_data.temp, data);
+  // metadata
+  Metadata meta = Metadata_init_zero;
+  meta.ts = ts.Seconds;
+  meta.logger_id = cfg->logger_id;
+  meta.cell_id = cfg->cell_id;
+
+  size_t data_len = 0;
+  SensorStatus sen_status = SENSOR_OK;
+
+  // matric potential
+  sen_status =
+      EncodeDoubleMeasurement(meta, sens_data.matric_pot,
+                              SensorType_TEROS21_MATRIC_POT, data, &data_len);
+  if (sen_status != SENSOR_OK) {
+    return -1;
+  }
+  SensorsAddMeasurement(data, data_len);
+
+  // temperature
+  sen_status = EncodeDoubleMeasurement(
+      meta, sens_data.temp, SensorType_TEROS21_TEMP, data, &data_len);
+  if (sen_status != SENSOR_OK) {
+    return -1;
+  }
 
   return data_len;
 }